TP1/hit8/requirements.txt
selenium==4.16.0
psycopg2-binary==2.9.9
